Backend/requirement.txt
langchain-core
langchain-groq
python-dotenv
speechrecognition
pyaudio
gtts
pygame
fastapi 
uvicorn
langchain